configs/avnet_rzboard_v2l: new defconfig
This patch adds support for the AVNET RZBoard V2L.

The board support creates a bootable sd card image. The board is
shipped with a u-boot inside its eMMC. The board support uses this
u-boot and does not build and deploy a u-boot onto the sd card.

Instead the sd card only contains a uEnv.txt, a kernel image and the
rootfs.

Signed-off-by: Kilian Zinnecker <kilian.zinnecker@mail.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>