package/fcft: disable uClibc support
fcft needs ISO C threads, aka C11 threads, which are different from
POSIX threads, and are available in both glibc and musl, but not in
uClibc-ng.

Add a dependency on glibc or musl, and since both imply threads, drop
the dependency on threads.

Propagate the new dependency to the foot package, and adapt the comments
accordingly.

Fixes:
http://autobuild.buildroot.net/results/1b2dda32a90d7db8117eede09f46d44cffe60993
http://autobuild.buildroot.net/results/5f03bb5c0c107a2b908be974421ae0a032d4dd89
http://autobuild.buildroot.net/results/31a1f022414a8a1888cac4a5cd1761257d0f2196
http://autobuild.buildroot.net/results/4522781e0448b325b54e93bbfb20f8295ea7061d

Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>
[yann.morin.1998@free.fr;
  - drop the dependencies on threads, implied by glibc or musl
  - fix dependencies on comments
  - reword the commit log to refer to ICO C threads
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>