package/qwt: reinstate mkspecs cleanup
The qwt package is a Qt5 package, not a Qt6 one, so it can't benefit
from the target-finalize hook that was introduced in e29fcde4ba70
(package/qt6base: remove $(TARGET_DIR)/usr/mkspecs in target-finalize).

Indeed, the original patch was hard-coding the mkspecs removal directly
in the target-finalize recipe, and that was made a hook as a last-minute
change when applying, without realising that qwt was qt5-based.

So, reinstate the mkspecs removal as an ad-hoc fixup in the qwt package.

This actually partially reverts commit e29fcde4ba70 (package/qt6base:
remove $(TARGET_DIR)/usr/mkspecs in target-finalize).

Reported-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>